// src/utils/types.ts
import { allSongsFlatWithGameSeries } from "../assets/songs";

export type SongType = (typeof allSongsFlatWithGameSeries)[number];
export type Results = Record<
  number,
  { loading: boolean; results?: TrackResultWithSimilarity[] }
>;

export type TrackResultWithSimilarity = {
  trackNameSimilarity: number;
  artistSimilarity: number;
} & TrackResult;

export type TrackResult = {
  album: {
    album_type: string;
    artists: {
      external_urls: {
        spotify: string;
      };
      href: string;
      id: string;
      name: string;
      type: string;
      uri: string;
    }[];
    available_markets: string[];
    external_urls: {
      spotify: string;
    };
    href: string;
    id: string;
    images: {
      height: number;
      width: number;
      url: string;
    }[];
    is_playable: boolean;
    name: string;
    release_date: string;
    release_date_precision: string;
    total_tracks: number;
    type: string;
    uri: string;
  };
  artists: {
    external_urls: {
      spotify: string;
    };
    href: string;
    id: string;
    name: string;
    type: string;
    uri: string;
  }[];
  available_markets: string[];
  disc_number: number;
  duration_ms: number;
  explicit: boolean;
  external_ids: {
    isrc: string;
  };
  external_urls: {
    spotify: string;
  };
  href: string;
  id: string;
  is_local: boolean;
  is_playable: boolean;
  name: string;
  popularity: number;
  preview_url: string | null;
  track_number: number;
  type: string;
  uri: string;
};

// export type TrackResult = {
//   name: string;
//   artists: {
//     name: string;
//     uri: string;
//   }[];
//   album: {
//     name: string;
//     images: {
//       url: string;
//     }[];
//   };
//   preview_url: string;
//   uri: string;
// };